+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Added Handler Sensu API HTTPS support through API configuration (e.g. `{"api": {"ssl": {}}`.
 
 ## [2.4.0] - 2018-02-08
 ### Added